clean up all services
