+++ app/lib/app.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:iwfpapp/screens/add_card_from_template/add_card_from_template_screen.dart';
 import 'package:iwfpapp/screens/add_promo/add_promo_screen.dart';
 import 'package:iwfpapp/screens/delete_account/delete_account_screen.dart';
+import 'package:iwfpapp/screens/edit_card/edit_card_screen.dart';
 import 'package:iwfpapp/screens/forgot_password/forgot_password_screen.dart';
 import 'package:iwfpapp/screens/home/main.dart';
 import 'package:iwfpapp/screens/placeholder/main.dart';
@@ -12,13 +13,10 @@ import 'package:iwfpapp/screens/screen_tester/main.dart';
 import 'package:iwfpapp/screens/sign_up/sign_up_screen.dart';
 import 'package:iwfpapp/screens/suggestion/main.dart';
 import 'package:iwfpapp/screens/add_card/add_card_screen.dart';
-import 'package:iwfpapp/screens/edit_card/main.dart';
-import 'package:iwfpapp/services/data_backend/base.dart';
 import 'package:iwfpapp/services/config/themes/default.dart';
 
 class IwfpApp extends StatelessWidget {
-  final DataBackend dataBackend;
-  const IwfpApp(this.dataBackend);
+  const IwfpApp();
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
