+++ app/lib/screens/edit_card/card_edited.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class CardEdited extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Text('done'),
+    );
+  }
+}
