+++ app/lib/screens/edit_card/edit_card_error.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class EditCardError extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Text('error'),
+    );
+  }
+}
