+++ app/lib/screens/edit_card/edit_card_screen.dart
@@ -1,0 +1,39 @@
+import 'package:flutter/material.dart';
+import 'package:iwfpapp/screens/edit_card/edit_card_content.dart';
+import 'package:iwfpapp/services/config/typedefs/credit_card.dart';
+import 'package:iwfpapp/services/data_backend/base.dart';
+import 'package:iwfpapp/widgets/layouts/preferred_width.dart';
+import 'package:provider/provider.dart';
+
+class EditCardScreen extends StatefulWidget {
+  @override
+  _EditCardScreen createState() {
+    return _EditCardScreen();
+  }
+}
+
+class _EditCardScreen extends State<EditCardScreen> {
+  CreditCard card = CreditCard('Unknown', 'unknown');
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    if (ModalRoute.of(context).settings.arguments != null) {
+      card = ModalRoute.of(context).settings.arguments;
+    }
+    Provider.of<DataBackend>(context, listen: false).maybeRefresh();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Edit Card'),
+      ),
+      body: PreferredWidthContent(
+          child: EditCardContent(
+        card: this.card,
+      )),
+    );
+  }
+}
