+++ app/lib/screens/edit_card/editing_card.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class EditingCard extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Center(
+        child: CircularProgressIndicator(),
+      ),
+    );
+  }
+}
